Appropriate model object default attributes. Closes #104.
Only three attributes are set to null out of lack of better default
(each given a comment to explain null assignment).
- User.homeLocation
- Claim.approver
- Item.geolocation

Item.receipt is set to a new Receipt object, though the Receipt
constructor just sets the fields to null. A TODO task is present in
Receipt to find a better default.