Added missing licenses. Put all licenses at very start of files.
There was very inconsistent license location (some before imports and
some after).  If a license is after the package and there are no
imports, then something is imported automatically, the import is placed
before the licence. To avoid this problem with files that currently have
no imports, and to ensure the license can be seen right away upon
opening the file in any editor, all of the licenses are now at the very
start of each file, before the package lines.